Adding contiuous integration and fixing mysql connection
Should have fixed mysql connection. Please look at the mysql_pool tests to get an idea of the syntax to use for the next ones.
SQL scripts were modified to help fix the issues. Pleas note that s-park_code is not intended to be used to import the database directly : its syntax was modified to ease the pasing of the queries to process them in the code.

Also, continuous integration seems to be functionning well.